refactor: Migrated DriveInfosManager to the Transactionable shared implementation
