+++ kDriveCore/Data/Cache/DriveInfosManager/DriveInfosManager+Transactionable.swift
@@ -1,0 +1,36 @@
+/*
+ Infomaniak kDrive - iOS App
+ Copyright (C) 2023 Infomaniak Network SA
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import Foundation
+import RealmSwift
+
+public extension DriveInfosManager {
+    func writeTransaction(withRealm realmClosure: (Realm) throws -> Void) throws {
+        try transactionExecutor.writeTransaction(withRealm: realmClosure)
+    }
+
+    func fetchObject<Element: Object>(ofType type: Element.Type,
+                                      withRealm realmClosure: (Realm) -> Element?) -> Element? {
+        return transactionExecutor.fetchObject(ofType: type, withRealm: realmClosure)
+    }
+
+    func fetchResults<Element: RealmFetchable>(ofType type: Element.Type,
+                                               withRealm realmClosure: (Realm) -> Results<Element>) -> Results<Element> {
+        return transactionExecutor.fetchResults(ofType: type, withRealm: realmClosure)
+    }
+}
